Hapus entities dan cukup gunakan class model saja
Penggunaan entity di sini sebenarnya sudah benar untuk projek kecil. Hanya saja best practice-nya nanti kalau projeknya sudah kompleks seharusnya property yang ada didalam class model dan entity harus berbeda. Karena entity di projek ini cuma dipakai untuk disimpan kedalam SharedPreferences dan property-nya juga sama dengan class model maka, file-file entity-nya saya hapus saja karena kesannya tidak ada gunanya si entity ini. Dan nanti kedepannya seharusnya class model dan entity itu extends ke Equatable. Contoh lebih detailnya bisa lihat di projek nusawork prospek ya.
